apps/trainer/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pixel-sheriff-trainer"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "redis>=5.2",
  "fastapi>=0.115",
  "uvicorn>=0.30",
  "numpy>=1.26",
  "torch>=2.0",
  "torchvision>=0.15",
  "pillow>=10.0",
  "onnx>=1.16",
  "onnxruntime>=1.18",
  "onnxscript>=0.1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[tool.pytest.ini_options]
pythonpath = ["src", "../../packages/pixel_sheriff_ml/src"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
